requirements.txt
django==3.1
Pillow==7.1
djangorestframework==3.12
psycopg2-binary==2.8.6
python-decouple
django-filter
drf-yasg